requirements.txt
flask 
flasgger
openai
azure-ai-documentintelligence>=1.0.0
pydantic
rich
python-dotenv